CVM-72: Update JUnit to version 4.12
Updated JUnit to version 4.12.  Version 4.12 is required due to upgrade
of Spring framework.  Added hamcrest-core JAR which is a dependency of
JUnit 4.12.

Removed junit JAR from 3rd party JAR lists in consumers and
sdt-consumer-sample build-extras.xml files.  This is only used for
testing so no need to include it when projects are built.

Added GRANT INHERIT PRIVILEGES ON USER SYSTEM TO SDT_OWNER to database
create_users.sql.  This is needed so that prepare_for_dbunit_load
procedure can be executed on Oracle12c.  Similar change was made for
MCOL in MME-147.

Changed DBUnitUtility getConnectionInstance() to set AutoCommit to
false.

Changed ojdbc JAR in userlibraries template file to ojdbc6.  This allows
tests to be run locally on Oracle12c database.