SDT-25
Updated subprojects configuration in build.gradle to remove the following dependencies:

* integration task test task dependency
* jacocoTestReport task integration task dependency

This ensures that the test task only runs the unit tests and not the unit tests and integration tests.

Added dependency on allprojects*.jacocoTestReport to codeCoverageReport task.  This prevents the following warning being displayed for each module when the task is executed "Reason: Task ':codeCoverageReport' uses this output of task ':<module_name>:jacocoTestReport' without declaring an explicit or implicit dependency."