SDT-25: Fix XSD validation unit tests
A bug in the openJDK implementation of xerces 2.12.2 causes cvc-complex-type.4 errors to be reported twice.  This causes two of the XSD validation unit tests to fail when using openJDK 11.0.17 or above.  To fix this added a dependency for the actual version of xerces 2.12.2 which doesn't have the bug in it.

Removed duplicate cvc-complex-type.4 errors from BulkSubmitRequestMandatoryMissingErrorMessages.txt and SubmitQueryRequestMandatoryMissingErrorMessages.txt.

Added suppression for CVE-2022-1471 which is flagged by Template CI dependency check stage.