Updated DependencyCheck version to 8.0.2
Updated DependencyCheck version to 8.0.2.  This is needed to fit in with a change to be made to the common pipeline to utilise the global suppressions file (see https://jeremylong.github.io/DependencyCheck/suppressions/publishedSuppressions.xml).

In addition, updated versions plugin to latest version (0.44.0) and removed duplicate suppressions from suppressions.xml.