SDT-138: Improve consumers module test coverage
Uncommented line in build.gradle that fails build if code coverage threshold isn't met.

Removed "test.finalizedBy jacocoTestReport" line from subprojects section of build.gradle.  This ran the jacocoTestReport task whenever any tests were run, even when running a subset of tests locally.  As a result of this the local run would result in a false coverage failure because not all of the tests in the module were being run.  The coverage checks are still carried out in a build due to the dependency on the check task.