SDT-201: Updates to classes used to convert XML to Json
First set of updates to classes used to convert a CBUS XML request to Json for passing on to CMC.  Changes fix the order of values in the Json and the property names so that they match the design.  Updated existing tests and created new tests to cover conversion of XML to request objects and conversion of request objects to Json.

Added timezone parameter to JsonFormat annotations that format dates so that they aren't affected by daylight saving time changes.

Removed CorrespondenceDetailType from request classes.  Correspondence address details are not included in information sent to CMC so no need to capture that info from the XML.

Removed convertXmlToJson() method from XmlConverter class.  Requests are never converted directly from XML to Json and method was only used in small number of tests, so it wasn't needed.  Also, the conversion to Json is now tested by the new test classes.

Removed redundant tests from XmlToObjectTest and added new tests.

IMPORTANT: There's an unresolved issue with the parsing of particulars lines from claim request XML.  This currently causes the XmlToObjectTest.shouldConvertClaimRequest() test to fail.  This will be addressed in the next update.