SDT-201: Handle timeouts from CMC
Updated create date methods used in CMC request/response tests to explicitly use Europe/London timezone.  This is to resolve an issue with mismatching timezones when the tests are run as part of the build pipelines.

Updated CMCConsumerGateway so that it doesn't try to convert the response detail to XML if it is null.