SDT-203: Handle CMC errors
Updated sdt-common version to 3.2.1-phase-3.0.5.  This is the "official" release of 3.2.1-SDT-200-2.0.1.

The release of sdt-common contains changes to allow a status of "Case Locked" to be returned in bulk feedback requests.  This should've been included in the SDT-200 changes, but was only noticed during testing of the STD-203 changes.